+++ backend/user_api/models.py
@@ -1,16 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import User
-
-
-
-class Sport(models.Model):
-    abrev = models.CharField(max_length=5, unique=True)
-    name = models.CharField(max_length=128, unique=True)
-    icon_url = models.URLField(max_length=128, blank=True)
-
-    def __str__(self):
-        return self.name
-    
+from sport_api.models import Sport   
 
 
 class AppUser(models.Model):
